+++ mne/decoding/tests/test_csp.py
@@ -40,18 +40,20 @@ def test_csp():
     y = epochs.events[:, -1]
 
     # Init
-    assert_raises(ValueError, CSP, n_components='foo')
+    assert_raises(ValueError, CSP, n_components='foo', norm_trace=False)
     for reg in ['foo', -0.1, 1.1]:
-        assert_raises(ValueError, CSP, reg=reg)
+        assert_raises(ValueError, CSP, reg=reg, norm_trace=False)
     for reg in ['oas', 'ledoit_wolf', 0, 0.5, 1.]:
-        CSP(reg=reg)
+        CSP(reg=reg, norm_trace=False)
     for cov_est in ['foo', None]:
-        assert_raises(ValueError, CSP, cov_est=cov_est)
+        assert_raises(ValueError, CSP, cov_est=cov_est, norm_trace=False)
+    assert_raises(ValueError, CSP, norm_trace='foo')
     for cov_est in ['concat', 'epoch']:
-        CSP(cov_est=cov_est)
+        CSP(cov_est=cov_est, norm_trace=False)
 
     n_components = 3
-    csp = CSP(n_components=n_components)
+    # XXX Rank deficient if norm_trace False
+    csp = CSP(n_components=n_components, norm_trace=True)
 
     # Fit
     csp.fit(epochs_data, epochs.events[:, -1])
@@ -81,7 +83,8 @@ def test_csp():
 
     # Test covariance estimation methods (results should be roughly equal)
     np.random.seed(0)
-    csp_epochs = CSP(cov_est="epoch")
+    # XXX Rank deficient if norm_trace == False
+    csp_epochs = CSP(cov_est="epoch", norm_trace=True)
     csp_epochs.fit(epochs_data, y)
     for attr in ('filters_', 'patterns_'):
         corr = np.corrcoef(getattr(csp, attr).ravel(),
@@ -97,7 +100,8 @@ def test_csp():
 
     n_channels = epochs_data.shape[1]
     for cov_est in ['concat', 'epoch']:
-        csp = CSP(n_components=n_components, cov_est=cov_est)
+        # XXX Rank deficient if norm_trace == False
+        csp = CSP(n_components=n_components, cov_est=cov_est, norm_trace=True)
         csp.fit(epochs_data, epochs.events[:, 2]).transform(epochs_data)
         assert_equal(len(csp._classes), 4)
         assert_array_equal(csp.filters_.shape, [n_channels, n_channels])
@@ -109,7 +113,8 @@ def test_csp():
     feature_shape = [len(epochs_data), n_components]
     X_trans = dict()
     for log in (None, True, False):
-        csp = CSP(n_components=n_components, log=log)
+        # XXX Rank deficient if norm_trace == False
+        csp = CSP(n_components=n_components, log=log, norm_trace=True)
         assert_true(csp.log is log)
         Xt = csp.fit_transform(epochs_data, epochs.events[:, 2])
         assert_array_equal(Xt.shape, feature_shape)
@@ -119,15 +124,20 @@ def test_csp():
     # Different normalization return different transform
     assert_true(np.sum((X_trans['True'] - X_trans['False']) ** 2) > 1.)
     # Check wrong inputs
-    assert_raises(ValueError, CSP, transform_into='average_power', log='foo')
+    assert_raises(ValueError, CSP, transform_into='average_power', log='foo',
+                  norm_trace=False)
 
     # Test csp space transform
-    csp = CSP(transform_into='csp_space')
+    csp = CSP(transform_into='csp_space', norm_trace=False)
     assert_true(csp.transform_into == 'csp_space')
     for log in ('foo', True, False):
-        assert_raises(ValueError, CSP, transform_into='csp_space', log=log)
+        assert_raises(ValueError, CSP, transform_into='csp_space', log=log,
+                      norm_trace=False)
+
     n_components = 2
-    csp = CSP(n_components=n_components, transform_into='csp_space')
+    # XXX Rank deficient if norm_trace == False
+    csp = CSP(n_components=n_components, transform_into='csp_space',
+              norm_trace=True)
     Xt = csp.fit(epochs_data, epochs.events[:, 2]).transform(epochs_data)
     feature_shape = [len(epochs_data), n_components, epochs_data.shape[2]]
     assert_array_equal(Xt.shape, feature_shape)
@@ -149,7 +159,7 @@ def test_regularized_csp():
     n_components = 3
     reg_cov = [None, 0.05, 'ledoit_wolf', 'oas']
     for reg in reg_cov:
-        csp = CSP(n_components=n_components, reg=reg)
+        csp = CSP(n_components=n_components, reg=reg, norm_trace=False)
         csp.fit(epochs_data, epochs.events[:, -1])
         y = epochs.events[:, -1]
         X = csp.fit_transform(epochs_data, y)
@@ -175,7 +185,7 @@ def test_csp_pipeline():
     """
     from sklearn.svm import SVC
     from sklearn.pipeline import Pipeline
-    csp = CSP(reg=1)
+    csp = CSP(reg=1, norm_trace=False)
     svc = SVC()
     pipe = Pipeline([("CSP", csp), ("SVC", svc)])
     pipe.set_params(CSP__reg=0.2)